refactor `cosid-benchmark`
