+++ cosid-benchmark/gradle.properties
@@ -12,4 +12,4 @@
 #
 
 group=me.ahoo.cosid.benchmark
-version=1.7.2
+version=1.9.0
